Add `startstop.StartAll` helper to start a series of services (#517)
Add a `StartAll` helper that starts a number of services at once, also
taking care that in the event of a problem, any services that had
started are once stopped. I found the use of something like this
convenient while trying some service code in River UI.

I also make a couple tweaks so that the `Stopped` channel becomes a
little safer to use. Like `Started`, it will preallocate itself if
called before a service is started. It's still not safe to call while
stopping or after a stop, but its ergonomics still measurably improve.

Lastly, I pull a somewhat unrelated change to add a `NewArchetype`
function in for easily creating an archetype in non-test environments.
Also found this useful when bringing use of archetypes into River UI.