Migrations: use a txn + commit for each migration, deprecate MigrateTx (#600)
This is partly extracted from #590. As detailed in that PR, there are certain
combinations of schema changes which are not allowed to be run within the same
transaction. The example we encountered with #590 is adding a new enum value,
then using it in an immutable function during a subsequent migration. In
Postgres, these must be separated by a commit.

The migration in that PR exposes an issue with our migration framework, which is
that there are certain schema changes which must be committed before they can be
referenced by subsequent schema changes. The example of this I found is that if
you add a new value to an enum, you cannot later create a function that relies
on that new enum value unless the new value has first been committed. Committing
it within a DDL transaction does _not_ count—it must be a full commit.

IMO this might mean that the entire idea of a `MigrateTx` API is not workable
with certain schema change combinations. At worst, it can result in
unpredictable failures depending on the exact sequence of changes and how many
migrations are being run at once.

As such, in this PR I've deprecated `MigrateTx` and adjusted the migrator so
that it opens a new transaction for each individual migration with a commit
between them. Migrator tests were changed to move away from `MigrateTx` and to a
setup where they get a new clean schema for each test that's disposed at the
end. This makes it possible to test the full sequence of database migrations
with a clean slate each time.

I believe this is the right long-term direction because it's the approach that
other migration libraries use (Rails/ActiveRecord, [Sequel][1], Goose, etc). It
also enables the potential for us to add the ability for individual migrations
to opt _out_ of a having a wrapping transaction, which is essential if we ever
want our default to be `CREATE INDEX CONCURRENTLY` rather than synchronous
indexing (as it really should be for any at-scale system).

[1]: (https://sequel.jeremyevans.net/rdoc/files/doc/migration_rdoc.html#label-Transactions)