Drop internalized random number generators in favor of `math/rand/v2` (#691)
The service archetype and a number of functions internalized a random
source because in Go, that used to be the only way that you could
guarantee that a non-crypto generated was actually seeded.

Go 1.22 introduced `math/rand/v2` which fixed that problem. Now that Go
1.21 support's been dropped, it's a nice clean up to move to the V2
package and get rid of random sources being passed around everywhere.